requirements.txt
typer==0.12.3
pillow==10.3.0
httpx==0.27.0
pydantic==2.8.2
pydantic-settings==2.3.4
python-dotenv==1.0.1
structlog==24.1.0
Babel==2.15.0
streamlit==1.37.0
pytest==8.2.1
ruff==0.5.7
black==24.8.0
pre-commit==3.8.0


pyyaml
openai>=1.30.0
